requirements.txt
playwright==1.40.0
google-generativeai==0.8.3
pillow==10.4.0
requests==2.31.0
python-dotenv>=1.0.1
flask==3.0.0
flask-socketio==5.3.6
